Adapt integration test to match wanted output
The final output differs a little from the initial idea because during
implementation I came to the conclusion that it would be better to also display
subdirectories (not only top level directories) if they are among the top 10.

The sizes and ordering of the top level directories and files matches those of
the original integration tests as can be seen in the diff.